Add an agentic analyst (instead of the hard-coded one)
This should be better in the long term, esp. when it comes to grading
runs (as it is able to detect compromised accounts, etc.)

Still need to wire up most things (e.g., the knowledge databases.. which
should also be included in multiple prompts). I also feel that the
agentic analyzer is currently performing worse than the old-schoool one..
but it's just a first shot.